Group Announcement cache-related work.
Group announcement counts (and title of the last announcement) are displayed on the group listing page. Added a sweeper for GroupAnnouncements that invalidates the cache of that listing on create, update & delete of a group announcement.

Caching group announcements themself is not possible, as the data is user-specific: showing only public of private and public announcements.

git-svn-id: http://myexperiment.rubyforge.org/svn/trunk@1813 4f7b4649-9ca9-420a-81b3-55a188e4d46f